+++ crates/core/src/fetch_manager.rs
@@ -1,10 +1,34 @@
 use std::{error, fs, path::PathBuf};
 
-use chromiumoxide::{BrowserConfig, BrowserFetcher, BrowserFetcherOptions};
+use chromiumoxide::{
+    BrowserConfig, BrowserFetcher, BrowserFetcherOptions,
+    detection::{DetectionOptions, default_executable},
+};
 use ffmpeg_sidecar::{
     command::FfmpegCommand,
     download::{download_ffmpeg_package, ffmpeg_download_url, unpack_ffmpeg},
 };
+
+pub fn is_chromium_installed() -> bool {
+    let cache_path = get_cache_path();
+    let download_path = cache_path.join("chromium");
+    if download_path.join("installed.txt").exists() {
+        return true;
+    }
+
+    default_executable(DetectionOptions::default()).is_ok()
+}
+pub fn is_ffmpeg_installed() -> bool {
+    let cache_path = get_cache_path();
+    let mut ffmpeg_path = cache_path.join("ffmpeg").join("ffmpeg");
+    if cfg!(windows) {
+        ffmpeg_path.set_extension("exe");
+    }
+    if ffmpeg_path.exists() {
+        return true;
+    }
+    ffmpeg_sidecar::command::ffmpeg_is_installed()
+}
 
 pub fn fetch_chromium() -> Result<(), Box<dyn error::Error>> {
     let rt = tokio::runtime::Builder::new_current_thread()